requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
Flask-PyMongo==2.3.0
pymongo==4.10.1
python-dotenv==1.0.1
bcrypt==4.2.0
email-validator==2.2.0
requests==2.32.3
pytest==7.4.2
black==23.9.1